# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TacoCloud Public API
  version: 1.0.0
  description: |
    Contrato HTTP público v1. Las rutas equivalentes bajo `/api` son aliases
    temporales y responden con `Deprecation` y `Link` hacia su sucesor v1.
    La versión HTTP es independiente de la versión del contrato de eventos.
servers:
  - url: /
...
security:
  - basicAuth: []
paths:
...
  /api/v1/tacos:
    get:
      tags: [Tacos]
      summary: Search the taco catalog
      security: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - {name: name, in: query, schema: {type: string, maxLength: 50}, example: veggie}
        - {name: ingredientId, in: query, schema: {type: string, maxLength: 20}, example: FLTO}
        - {name: diet, in: query, schema: {$ref: '#/components/schemas/DietaryTag'}}
        - {name: excludeAllergen, in: query, schema: {$ref: '#/components/schemas/Allergen'}}
        - {name: spice, in: query, schema: {$ref: '#/components/schemas/SpiceLevel'}}
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/Size'}
        - name: sort
          in: query
          schema:
            type: string
            enum: ['createdAt,desc', 'createdAt,asc', 'name,desc', 'name,asc', 'id,desc', 'id,asc']
          example: createdAt,desc
      responses:
        '200':
          description: Paged taco catalog
          headers:
            X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TacoSearchResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '422': {$ref: '#/components/responses/Unprocessable'}
    post:
      tags: [Tacos]
      summary: Publish a taco design
      description: Requires USER or ADMIN. Only name and ingredient references are accepted contractually.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TacoCreateRequest'}
      responses:
        '201':
          description: Taco created
          headers:
            X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TacoCatalogResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '422': {$ref: '#/components/responses/Unprocessable'}
  /api/v1/tacos/today:
    get:
      tags: [Tacos]
      summary: Get the deterministic taco of the day
      security: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Daily recommendation
          headers:
            X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DailyTacoResponse'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/tacos/validate:
    post:
      tags: [Tacos]
      summary: Validate a taco design without saving it
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TacoDesignRequest'}
      responses:
        '200':
          description: Validation result
          headers:
            X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TacoDesignValidationResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /api/v1/tacos/top:
    get:
      tags: [Ratings]
      summary: Get top-rated tacos
      security: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 50, default: 10}
          example: 10
      responses:
        '200':
          description: Aggregated ranking
          headers:
            X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/TopTacoRatingResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
  /api/v1/tacos/{id}:
    parameters:
      - $ref: '#/components/parameters/TacoId'
      - $ref: '#/components/parameters/CorrelationId'
    get:
      tags: [Tacos]
      summary: Get a taco
      security: []
      responses:
        '200':
          description: Taco found
          headers:
            X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TacoCatalogResponse'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/tacos/{id}/classification:
    parameters:
      - $ref: '#/components/parameters/TacoId'
      - $ref: '#/components/parameters/CorrelationId'
    get:
      tags: [Tacos]
      summary: Get derived dietary classification
      security: []
      responses:
        '200':
          description: Taco classification
          headers:
            X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TacoClassificationResponse'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/tacos/{id}/rating:
    parameters:
      - $ref: '#/components/parameters/TacoId'
      - $ref: '#/components/parameters/CorrelationId'
    put:
      tags: [Ratings]
      summary: Set the authenticated user's rating
      description: Requires USER or ADMIN; repeating PUT replaces that user's score.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RatingRequest'}
      responses:
        '204': {$ref: '#/components/responses/NoContent'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: HTTP Basic credentials resolved by Spring Security. Browser form login may establish the same authenticated session.
  headers:
    CorrelationId:
      description: Validated request correlation or server-generated UUID.
      schema:
        type: string
        maxLength: 64
        pattern: '^[A-Za-z0-9._-]+$'
      example: client-operation-123
  parameters:
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      description: Preserved when valid; otherwise the server generates a UUID and returns it in the response.
      schema:
        type: string
        maxLength: 64
        pattern: '^[A-Za-z0-9._-]+$'
      example: client-operation-123
...
    TacoId:
      name: id
      in: path
      required: true
      schema: {type: string}
...
    Page:
      name: page
      in: query
      schema: {type: integer, minimum: 0, default: 0}
    Size:
      name: size
      in: query
      schema: {type: integer, minimum: 1, maximum: 50, default: 20}
  responses:
    NoContent:
      description: Operation completed without a response body
      headers:
        X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
    BadRequest:
      description: Invalid request or malformed JSON
      headers:
        X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/ApiProblem'}
          example:
            type: urn:tacocloud:problem:validation_error
            title: Invalid request
            status: 400
            detail: The request contains invalid fields.
            instance: /api/v1/orders
            code: VALIDATION_ERROR
            violations:
              - field: deliveryName
                reason: deliveryName is required
    Unauthorized:
      description: Authentication is required
      headers:
        X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/ApiProblem'}
    Forbidden:
      description: Authenticated principal lacks the required role or ownership
      headers:
        X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/ApiProblem'}
    NotFound:
      description: Resource not found or hidden by ownership rules
      headers:
        X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/ApiProblem'}
...
    Unprocessable:
      description: Request is syntactically valid but violates a business rule
      headers:
        X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/ApiProblem'}
  schemas:
    IngredientType:
      type: string
      enum: [WRAP, PROTEIN, VEGGIES, CHEESE, SAUCE]
    DietaryTag:
      type: string
      enum: [VEGAN, VEGETARIAN, GLUTEN_FREE]
    Allergen:
      type: string
      enum: [GLUTEN, DAIRY, EGG, PEANUT, TREE_NUT, SOY]
    SpiceLevel:
      type: string
      enum: [NONE, MILD, MEDIUM, HOT, EXTREME]
...
    IngredientResponse:
      type: object
      additionalProperties: false
      required: [id, name, type, unitPrice, available, dietaryTags, allergens, spiceLevel]
      properties:
        id: {type: string}
        name: {type: string}
        type: {$ref: '#/components/schemas/IngredientType'}
        unitPrice: {type: number, format: decimal, minimum: 0, multipleOf: 0.01}
        available: {type: boolean}
        dietaryTags:
          type: array
          items: {$ref: '#/components/schemas/DietaryTag'}
        allergens:
          type: array
          items: {$ref: '#/components/schemas/Allergen'}
        spiceLevel: {$ref: '#/components/schemas/SpiceLevel'}
...
    IngredientReference:
      type: object
      additionalProperties: false
      required: [id]
      properties:
        id: {type: string, example: FLTO}
    TacoCreateRequest:
      type: object
      additionalProperties: false
      required: [name, ingredients]
      properties:
        name: {type: string, minLength: 1, maxLength: 50, example: Veggie Taco}
        ingredients:
          type: array
          minItems: 1
          maxItems: 50
          items: {$ref: '#/components/schemas/IngredientReference'}
    TacoDesignRequest:
      type: object
      additionalProperties: false
      required: [name, ingredientIds]
      properties:
        name: {type: string, minLength: 1, maxLength: 50}
        ingredientIds:
          type: array
          maxItems: 50
          items: {type: string, minLength: 1}
    TacoDesignViolationResponse:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code: {type: string}
        message: {type: string}
    TacoDesignValidationResponse:
      type: object
      additionalProperties: false
      required: [valid, violations]
      properties:
        valid: {type: boolean}
        violations:
          type: array
          items: {$ref: '#/components/schemas/TacoDesignViolationResponse'}
    TacoClassificationResponse:
      type: object
      additionalProperties: false
      required: [tacoId, tacoName, dietaryTags, allergens, spiceLevel, disclaimer]
      properties:
        tacoId: {type: string, nullable: true}
        tacoName: {type: string}
        dietaryTags:
          type: array
          items: {$ref: '#/components/schemas/DietaryTag'}
        allergens:
          type: array
          items: {$ref: '#/components/schemas/Allergen'}
        spiceLevel: {$ref: '#/components/schemas/SpiceLevel'}
        disclaimer: {type: string}
    TacoCatalogResponse:
      type: object
      additionalProperties: false
      required: [id, name, createdAt, ingredients, classification]
      properties:
        id: {type: string, nullable: true}
        name: {type: string}
        createdAt: {type: string, format: date-time, nullable: true}
        ingredients:
          type: array
          items: {$ref: '#/components/schemas/IngredientResponse'}
        classification: {$ref: '#/components/schemas/TacoClassificationResponse'}
    TacoSearchResponse:
      type: object
      additionalProperties: false
      required: [items, page, size, totalElements, totalPages]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/TacoCatalogResponse'}
        page: {type: integer, minimum: 0}
        size: {type: integer, minimum: 1}
        totalElements: {type: integer, format: int64, minimum: 0}
        totalPages: {type: integer, minimum: 0}
    DailyTacoResponse:
      type: object
      additionalProperties: false
      required: [taco, date, reason]
      properties:
        taco: {$ref: '#/components/schemas/TacoCatalogResponse'}
        date: {type: string, format: date}
        reason: {type: string}
    RatingRequest:
      type: object
      additionalProperties: false
      required: [score]
      properties:
        score: {type: integer, minimum: 1, maximum: 5}
    RatingDistribution:
      type: object
      additionalProperties:
        type: integer
        format: int64
        minimum: 0
      example: {'1': 0, '2': 0, '3': 1, '4': 2, '5': 4}
    TopTacoRatingResponse:
      type: object
      additionalProperties: false
      required: [taco, average, count, distribution]
      properties:
        taco: {$ref: '#/components/schemas/TacoCatalogResponse'}
        average: {type: number, format: decimal, minimum: 1, maximum: 5}
        count: {type: integer, format: int64, minimum: 0}
        distribution: {$ref: '#/components/schemas/RatingDistribution'}
...
    ApiProblem:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, instance, code, violations]
      properties:
        type: {type: string, format: uri}
        title: {type: string}
        status: {type: integer, minimum: 400, maximum: 599}
        detail: {type: string}
        instance: {type: string}
        code: {type: string}
        violations:
          type: array
          items: {$ref: '#/components/schemas/Violation'}
    Violation:
      type: object
      additionalProperties: false
      required: [field, reason]
      properties:
        field: {type: string}
        reason: {type: string}